+++ src/win32Main/kotlin/ingameIME/context/inputState/inputMode/conversion/jap/HiraganaMode.kt
@@ -26,7 +26,7 @@ object HiraganaMode : IHiraganaMode, ITfConversionMode {
      * Check whether the current mode contains itself
      */
     override fun check(mode: libtf_ConversionMode): Boolean {
-        return mode.toInt() and (TF_CONVERSIONMODE_KATAKANA or TF_CONVERSIONMODE_ROMAN) == 0
+        return mode and (TF_CONVERSIONMODE_KATAKANA or TF_CONVERSIONMODE_ROMAN) == 0
     }
 
     override fun onApplyState(context: IInputContext) {
@@ -35,7 +35,7 @@ object HiraganaMode : IHiraganaMode, ITfConversionMode {
                 val mode: libtf_ConversionModeVar = this.alloc()
                 libtf_get_conversion_mode(it.nativeContext.value, mode.ptr).succeedOrThr()
                 mode.value =
-                    mode.value and TF_CONVERSIONMODE_KATAKANA.toUInt().inv() and TF_CONVERSIONMODE_ROMAN.toUInt().inv()
+                    mode.value and TF_CONVERSIONMODE_KATAKANA.inv() and TF_CONVERSIONMODE_ROMAN.inv()
                 libtf_set_conversion_mode(it.nativeContext.value, mode.value).succeedOrThr()
             }
         }
