+++ src/win32Main/kotlin/ingameIME/context/inputState/inputMode/conversion/jap/KatakanaMode.kt
@@ -27,7 +27,7 @@ object KatakanaMode : IKatakanaMode, ITfConversionMode {
      * Check whether the current mode contains itself
      */
     override fun check(mode: libtf_ConversionMode): Boolean {
-        return mode.toInt() and TF_CONVERSIONMODE_KATAKANA == 1
+        return mode and TF_CONVERSIONMODE_KATAKANA == 1
     }
 
     override fun onApplyState(context: IInputContext) {
@@ -36,7 +36,7 @@ object KatakanaMode : IKatakanaMode, ITfConversionMode {
                 val mode: libtf_ConversionModeVar = this.alloc()
                 libtf_get_conversion_mode(it.nativeContext.value, mode.ptr).succeedOrThr()
                 mode.value =
-                    mode.value and TF_CONVERSIONMODE_ROMAN.toUInt().inv() or TF_CONVERSIONMODE_KATAKANA.toUInt()
+                    mode.value and TF_CONVERSIONMODE_ROMAN.inv() or TF_CONVERSIONMODE_KATAKANA
                 libtf_set_conversion_mode(it.nativeContext.value, mode.value).succeedOrThr()
             }
         }
